requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.1.7
django-storages==1.13.2
filelock==3.9.0
idna==3.4
pipenv==2023.2.18
platformdirs==3.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyAudio==0.2.13
pycparser==2.20
pyobjc==9.0.1
pyobjc-core==9.0.1
pyobjc-framework-Accessibility==9.0.1
pyobjc-framework-Accounts==9.0.1
pyobjc-framework-AddressBook==9.0.1
pyobjc-framework-AdServices==9.0.1
pyobjc-framework-AdSupport==9.0.1
pyobjc-framework-AppleScriptKit==9.0.1
pyobjc-framework-AppleScriptObjC==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-AppTrackingTransparency==9.0.1
pyobjc-framework-AudioVideoBridging==9.0.1
pyobjc-framework-AuthenticationServices==9.0.1
pyobjc-framework-AutomaticAssessmentConfiguration==9.0.1
pyobjc-framework-Automator==9.0.1
pyobjc-framework-AVFoundation==9.0.1
pyobjc-framework-AVKit==9.0.1
pyobjc-framework-AVRouting==9.0.1
pyobjc-framework-BackgroundAssets==9.0.1
pyobjc-framework-BusinessChat==9.0.1
pyobjc-framework-CalendarStore==9.0.1
pyobjc-framework-CallKit==9.0.1
pyobjc-framework-CFNetwork==9.0.1
pyobjc-framework-ClassKit==9.0.1
pyobjc-framework-CloudKit==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Collaboration==9.0.1
pyobjc-framework-ColorSync==9.0.1
pyobjc-framework-Contacts==9.0.1
pyobjc-framework-ContactsUI==9.0.1
pyobjc-framework-CoreAudio==9.0.1
pyobjc-framework-CoreAudioKit==9.0.1
pyobjc-framework-CoreBluetooth==9.0.1
pyobjc-framework-CoreData==9.0.1
pyobjc-framework-CoreHaptics==9.0.1
pyobjc-framework-CoreLocation==9.0.1
pyobjc-framework-CoreMedia==9.0.1
pyobjc-framework-CoreMediaIO==9.0.1
pyobjc-framework-CoreMIDI==9.0.1
pyobjc-framework-CoreML==9.0.1
pyobjc-framework-CoreMotion==9.0.1
pyobjc-framework-CoreServices==9.0.1
pyobjc-framework-CoreSpotlight==9.0.1
pyobjc-framework-CoreText==9.0.1
pyobjc-framework-CoreWLAN==9.0.1
pyobjc-framework-CryptoTokenKit==9.0.1
pyobjc-framework-DataDetection==9.0.1
pyobjc-framework-DeviceCheck==9.0.1
pyobjc-framework-DictionaryServices==9.0.1
pyobjc-framework-DiscRecording==9.0.1
pyobjc-framework-DiscRecordingUI==9.0.1
pyobjc-framework-DiskArbitration==9.0.1
pyobjc-framework-DVDPlayback==9.0.1
pyobjc-framework-EventKit==9.0.1
pyobjc-framework-ExceptionHandling==9.0.1
pyobjc-framework-ExecutionPolicy==9.0.1
pyobjc-framework-ExtensionKit==9.0.1
pyobjc-framework-ExternalAccessory==9.0.1
pyobjc-framework-FileProvider==9.0.1
pyobjc-framework-FileProviderUI==9.0.1
pyobjc-framework-FinderSync==9.0.1
pyobjc-framework-FSEvents==9.0.1
pyobjc-framework-GameCenter==9.0.1
pyobjc-framework-GameController==9.0.1
pyobjc-framework-GameKit==9.0.1
pyobjc-framework-GameplayKit==9.0.1
pyobjc-framework-HealthKit==9.0.1
pyobjc-framework-ImageCaptureCore==9.0.1
pyobjc-framework-IMServicePlugIn==9.0.1
pyobjc-framework-InputMethodKit==9.0.1
pyobjc-framework-InstallerPlugins==9.0.1
pyobjc-framework-InstantMessage==9.0.1
pyobjc-framework-Intents==9.0.1
pyobjc-framework-IntentsUI==9.0.1
pyobjc-framework-IOSurface==9.0.1
pyobjc-framework-iTunesLibrary==9.0.1
pyobjc-framework-KernelManagement==9.0.1
pyobjc-framework-LatentSemanticMapping==9.0.1
pyobjc-framework-LaunchServices==9.0.1
pyobjc-framework-libdispatch==9.0.1
pyobjc-framework-LinkPresentation==9.0.1
pyobjc-framework-LocalAuthentication==9.0.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==9.0.1
pyobjc-framework-MailKit==9.0.1
pyobjc-framework-MapKit==9.0.1
pyobjc-framework-MediaAccessibility==9.0.1
pyobjc-framework-MediaLibrary==9.0.1
pyobjc-framework-MediaPlayer==9.0.1
pyobjc-framework-MediaToolbox==9.0.1
pyobjc-framework-Metal==9.0.1
pyobjc-framework-MetalFX==9.0.1
pyobjc-framework-MetalKit==9.0.1
pyobjc-framework-MetalPerformanceShaders==9.0.1
pyobjc-framework-MetalPerformanceShadersGraph==9.0.1
pyobjc-framework-MetricKit==9.0.1
pyobjc-framework-MLCompute==9.0.1
pyobjc-framework-ModelIO==9.0.1
pyobjc-framework-MultipeerConnectivity==9.0.1
pyobjc-framework-NaturalLanguage==9.0.1
pyobjc-framework-NetFS==9.0.1
pyobjc-framework-Network==9.0.1
pyobjc-framework-NetworkExtension==9.0.1
pyobjc-framework-NotificationCenter==9.0.1
pyobjc-framework-OpenDirectory==9.0.1
pyobjc-framework-OSAKit==9.0.1
pyobjc-framework-OSLog==9.0.1
pyobjc-framework-PassKit==9.0.1
pyobjc-framework-PencilKit==9.0.1
pyobjc-framework-Photos==9.0.1
pyobjc-framework-PhotosUI==9.0.1
pyobjc-framework-PreferencePanes==9.0.1
pyobjc-framework-PushKit==9.0.1
pyobjc-framework-Quartz==9.0.1
pyobjc-framework-QuickLookThumbnailing==9.0.1
pyobjc-framework-ReplayKit==9.0.1
pyobjc-framework-SafariServices==9.0.1
pyobjc-framework-SafetyKit==9.0.1
pyobjc-framework-SceneKit==9.0.1
pyobjc-framework-ScreenCaptureKit==9.0.1
pyobjc-framework-ScreenSaver==9.0.1
pyobjc-framework-ScreenTime==9.0.1
pyobjc-framework-ScriptingBridge==9.0.1
pyobjc-framework-SearchKit==9.0.1
pyobjc-framework-Security==9.0.1
pyobjc-framework-SecurityFoundation==9.0.1
pyobjc-framework-SecurityInterface==9.0.1
pyobjc-framework-ServiceManagement==9.0.1
pyobjc-framework-SharedWithYou==9.0.1
pyobjc-framework-SharedWithYouCore==9.0.1
pyobjc-framework-ShazamKit==9.0.1
pyobjc-framework-Social==9.0.1
pyobjc-framework-SoundAnalysis==9.0.1
pyobjc-framework-Speech==9.0.1
pyobjc-framework-SpriteKit==9.0.1
pyobjc-framework-StoreKit==9.0.1
pyobjc-framework-SyncServices==9.0.1
pyobjc-framework-SystemConfiguration==9.0.1
pyobjc-framework-SystemExtensions==9.0.1
pyobjc-framework-ThreadNetwork==9.0.1
pyobjc-framework-UniformTypeIdentifiers==9.0.1
pyobjc-framework-UserNotifications==9.0.1
pyobjc-framework-UserNotificationsUI==9.0.1
pyobjc-framework-VideoSubscriberAccount==9.0.1
pyobjc-framework-VideoToolbox==9.0.1
pyobjc-framework-Virtualization==9.0.1
pyobjc-framework-Vision==9.0.1
pyobjc-framework-WebKit==9.0.1
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2021.1
requests==2.28.2
SpeechRecognition==3.9.0
speedtest-cli==2.1.3
sqlparse==0.4.3
urllib3==1.26.14
virtualenv==20.20.0
virtualenv-clone==0.5.7
